Add files via upload
This .py file reads and organizes data into a dataframe. It also changes data into binary elements, such that the data can be more easily processed by a machine learning algorithm.
Purpose: this dataframe is (almost) ideally how the data will be structured when we process it through the a machine learning algorithm. We should try to implement one of the machine learning algorithm templates into this .py file to see if they'll work.
How to use: download the file, or copy and paste. Only thing that needs changing are the two "path" variables, which should be the paths to where one stores there domoticz and ifttt .json files